clean up toast checks and organize event helpers
