+++ packages/manager/cypress/integration/linodes/smoke-delete-linode.spec.js
@@ -1,4 +1,3 @@
-import '@testing-library/cypress/add-commands';
 import { createLinode } from '../../support/api/linodes';
 import { getLinodeLandingRow } from '../../support/ui/linodes';
