+++ packages/manager/cypress/support/ui/events.js
@@ -1,0 +1,6 @@
+export const checkToast = message => {
+  cy.get('[data-qa-toast]').within(
+    e => cy.findByText(message, { exact: false }).should('be.visible')
+    // and should have class / success error
+  );
+};
